add tests for company and user signup
